added rounded corners transform for retweet user icon
